+++ task.go
@@ -87,7 +87,7 @@ func (p *EasyTask) Run() (ret []any, retErr error) {
 	// Start workers
 	for i := 0; i < p.config.worker; i++ {
 		opt := &worker{
-			id:         i,
+			id:         int64(i),
 			fun:        p.config.fun,
 			cmd:        make(chan msg),
 			resultCh:   resultCh,
